requirements.txt
geopy==2.1.0
pyserial==3.5 
pynmea2==1.15.0
matplotlib == 3.4.3
haversine==2.3.0
vincenty==0.1.4
numpy==1.21.1
statistics==1.0.3.5
PySimpleGui == 4.59.0 #comment this out when running on rpi
CMake
wmm2020
ahrs
opencv-python
#TODO: april tag library



 #import only if on rpi: 
# sparkfun-ublox-gps